moves validation logic for section inside entity class, moves add section hook to section entity file
